LumenGG/requirements.txt
asgiref==3.11.1
bleach==6.2.0
certifi==2024.12.14
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.1
colorama==0.4.6
daphne==4.2.1
Django==5.1.4
django-summernote==0.8.20.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
Markdown==3.10.2
martor==1.8.1
mysqlclient==2.2.8
openpyxl==3.1.5
packaging==24.2
pillow==12.2.0
qrcode[pil]==8.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
zipp==3.23.1
